requirements.txt
gradio
assemblyai
python-dotenv
ffmpeg-python
moviepy==1.0.3
yt-dlp>=2023.11.14
python-dotenv>=1.0.0
load_dotenv